+++ src/linkedListScreen.cpp
@@ -265,7 +265,7 @@ void LinkedListScreen::update(sf::RenderWindow& window, sf::Time deltaTime) {
         ImGui::Spacing();
 
         for (size_t i = 0; i < code.size(); ++i) {
-            if (i == activeLine) {
+            if ((int)i == activeLine) {
                 // Highlight the active line in yellow
                 ImGui::TextColored(ImVec4(1.0f, 1.0f, 0.0f, 1.0f), "%s", code[i].c_str());
             }
